move fontdata_namecheck out of googlefonts directory
